Use BigIntegerField for RandomNumber max value
This will allow us to accept higher numbers, we cannot write tests
because the types are not validated in SQLite.

+++ eas/api/models.py
@@ -149,7 +149,7 @@ class MultiResultMixin(models.Model):
 
 class RandomNumber(MultiResultMixin, BaseDraw):
     range_min = models.IntegerField()
-    range_max = models.IntegerField()
+    range_max = models.BigIntegerField()
 
     def generate_result_item(self):
         return random.randint(self.range_min, self.range_max)
